backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
sqlalchemy==2.0.27
asyncpg==0.29.0
alembic==1.13.1
python-dotenv==1.0.1
python-telegram-bot==20.7
